pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "adaptive-contrastive-curriculum-for-multitask-knowledge-transfer"
version = "0.1.0"
description = "Adaptive curriculum learning with contrastive uncertainty gradients for multi-task MMLU evaluation"
authors = [{name = "Alireza Shojaei"}]
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "datasets>=2.14.0",
    "sentence-transformers>=2.2.0",
    "evaluate>=0.4.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "mlflow>=2.5.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "scipy>=1.11.0",
    "sentencepiece>=0.1.99",
    "protobuf>=3.20.0",
    "tiktoken>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "flake8>=6.1.0",
    "mypy>=1.5.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--verbose --cov=src --cov-report=html --cov-report=term"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
